+++ config/symbols.pspeu.strno3.txt
@@ -33,10 +33,10 @@ EntityOrobourous = 0x09251DC8;
 EntityOrobSegment = 0x09252968;
 EntityOrobHeadParts = 0x09252CE8;
 EntityOrobRider = 0x09252E40;
-func_us_801C4334 = 0x092530B8;
+EntityOrobHolyAttacker = 0x092530B8;
 EntityDragonRider = 0x09253258;
-func_us_801C48D8 = 0x09253868;
-func_us_801C4B44 = 0x09253BD0;
+EntityDragonSegment = 0x09253868;
+EntityDragonHeadParts = 0x09253BD0;
 func_us_801C4C50 = 0x09253D28;
 RNO3_spriteBanks = 0x09263B60;
 RNO3_rooms_layers = 0x09263EB0;
