+++ src/st/rno3/rno3.h
@@ -59,11 +59,11 @@ typedef enum EntityIDs {
     /* 0x2B */ E_OROBOUROUS,             // EntityOrobourous
     /* 0x2C */ E_OROB_SEGMENT,           // EntityOrobSegment
     /* 0x2D */ E_OROB_HEAD_PARTS,        // EntityOrobHeadParts
-    /* 0x2E */ E_OROB_RIDER,                 // EntityOrobRider
-    /* 0x2F */ E_UNK_2F,                 // func_us_801C4334
+    /* 0x2E */ E_OROB_RIDER,             // EntityOrobRider
+    /* 0x2F */ E_OROB_HOLYATTACK,        // EntityOrobHolyAttacker
     /* 0x30 */ E_DRAGON_RIDER,           // EntityDragonRider
-    /* 0x31 */ E_UNK_31,                 // func_us_801C48D8
-    /* 0x32 */ E_UNK_32,                 // func_us_801C4B44
+    /* 0x31 */ E_DRAGON_SEGMENT,         // EntityDragonSegment
+    /* 0x32 */ E_DRAGON_HEAD_PARTS,      // EntityDragonHeadParts
     /* 0x33 */ E_UNK_33,                 // func_us_801C4C50
     /* 0x34 */ E_VENUS_WEED,             // EntityVenusWeed
     /* 0x35 */ E_UNK_35,                 // func_us_801C5850
