+++ src/st/rno3_psp/st_init.c
@@ -37,10 +37,10 @@ s32 E_ID(OROBOUROUS);
 s32 E_ID(OROB_SEGMENT);
 s32 E_ID(OROB_HEAD_PARTS);
 s32 E_ID(OROB_RIDER);
-s32 E_ID(UNK_2F);
+s32 E_ID(OROB_HOLYATTACK);
 s32 E_ID(DRAGON_RIDER);
-s32 E_ID(UNK_31);
-s32 E_ID(UNK_32);
+s32 E_ID(DRAGON_SEGMENT);
+s32 E_ID(DRAGON_HEAD_PARTS);
 s32 E_ID(UNK_33);
 s32 E_ID(VENUS_WEED);
 s32 E_ID(UNK_35);
@@ -101,10 +101,10 @@ void InitEntityIds(void) {
     E_ID(OROB_SEGMENT) = E_OROB_SEGMENT;
     E_ID(OROB_HEAD_PARTS) = E_OROB_HEAD_PARTS;
     E_ID(OROB_RIDER) = E_OROB_RIDER;
-    E_ID(UNK_2F) = E_UNK_2F;
+    E_ID(OROB_HOLYATTACK) = E_OROB_HOLYATTACK;
     E_ID(DRAGON_RIDER) = E_DRAGON_RIDER;
-    E_ID(UNK_31) = E_UNK_31;
-    E_ID(UNK_32) = E_UNK_32;
+    E_ID(DRAGON_SEGMENT) = E_DRAGON_SEGMENT;
+    E_ID(DRAGON_HEAD_PARTS) = E_DRAGON_HEAD_PARTS;
     E_ID(UNK_33) = E_UNK_33;
     E_ID(VENUS_WEED) = E_VENUS_WEED;
     E_ID(UNK_35) = E_UNK_35;
